Create cellangle.m
Could be useful? Also, it would be great if any of your cell operation routines accepted a scalar argument instead of one of the two cell argument. As in celladd(X,2) would add 2 to each element of the cell X. In this way the range of operations on cell would be infinite. In fact, you could write a routine called cellop(X,Y,...,'function') where function is a matlab function accepting matrix arguments X,Y etc